// src/components/molecules/datepicker/components/datepicker-orchestrator/components/datepicker-daypicker/components/datepicker-days/components/datepicker-day/enums/datepicker-day.component.enums.ts
enum DatepickerDayComponentStatesEnum {
  DEFAULT = 'default',
  DISABLED = 'disabled',
  DISABLED_TODAY = 'disabled-today',
  OUTSIDE = 'outside',
  OUTSIDE_SELECTED = 'outside-selected',
  OUTSIDE_TODAY = 'outside-today',
  SELECTED = 'selected',
  TODAY = 'today',
}

export { DatepickerDayComponentStatesEnum };